requirements.txt
selenium
gspread
google-auth
python-dotenv
webdriver-manager